feat: Add DarkLightImageAutoscale component for responsive logo styling
